深度优化了LQR、抗劈叉、滚转角、航向角跟随和腿长控制器的参数，大幅提高了机器人的运动性能和稳定性。 经测试，机器人可以通过标准的RM对抗赛场地的飞坡路段，且在飞坡过程中保持稳定。 经测试，机器人保持低姿态时可全速（约2m/s）通过RM对抗赛场地的搓板路段，且在过程中保持稳定。

+++ Chassis_Code_MC01/Task/inc/Communicate_Task.h
@@ -4,7 +4,7 @@
 typedef struct              
 {  
     int16_t Vx;	
-	int16_t Vx_step;//移动速度设定值的斜坡值
+	float Vx_step;//移动速度设定值的斜坡值
 	int16_t Yaw_Position_Move;
 	float SpeedZoom;//移动速度缩放系数，用于调整移速设定值的数值范围
 	uint8_t Chassis_Power_Switch;//底盘开关控制位
